requirements.txt
flask
sqlalchemy
apscheduler
pandas
statsmodels
requests
python-telegram-bot==13.15
